extract aio and reporter modules
